+++ src/redteam/index.ts
@@ -5,7 +5,7 @@ import logger from '../logger';
 import { loadApiProvider } from '../providers';
 import type { ApiProvider, TestCase } from '../types';
 import { extractVariablesFromTemplates } from '../util/templates';
-import { REDTEAM_MODEL, ALL_PLUGINS } from './constants';
+import { REDTEAM_MODEL, ALL_PLUGINS, HARM_CATEGORIES, PII_REQUEST_CATEGORIES } from './constants';
 import { addInjections } from './methods/injections';
 import { addIterativeJailbreaks } from './methods/iterative';
 import CompetitorPlugin from './plugins/competitors';
@@ -13,10 +13,10 @@ import ContractPlugin from './plugins/contracts';
 import DebugAccessPlugin from './plugins/debugInterface';
 import ExcessiveAgencyPlugin from './plugins/excessiveAgency';
 import HallucinationPlugin from './plugins/hallucination';
-import { HARM_CATEGORIES, getHarmfulTests } from './plugins/harmful';
+import { getHarmfulTests } from './plugins/harmful';
 import HijackingPlugin from './plugins/hijacking';
 import OverreliancePlugin from './plugins/overreliance';
-import { PII_REQUEST_CATEGORIES, getPiiLeakTestsForCategory } from './plugins/pii';
+import { getPiiLeakTestsForCategory } from './plugins/pii';
 import PoliticsPlugin from './plugins/politics';
 import RbacPlugin from './plugins/rbac';
 import ShellInjectionPlugin from './plugins/shellInjection';
